hw07: jdbc refactoring and fix tests

+++ hw07-JDBC/src/main/java/ru/otus/console/UserCommand.java
@@ -4,14 +4,13 @@ import org.apache.commons.lang3.StringUtils;
 import org.springframework.shell.standard.ShellComponent;
 import org.springframework.shell.standard.ShellMethod;
 import org.springframework.shell.standard.ShellOption;
-import ru.otus.entity.Book;
+import org.springframework.util.CollectionUtils;
 import ru.otus.service.BookService;
-
-import java.util.Collections;
-import java.util.List;
 
 @ShellComponent
 public class UserCommand {
+    private static final String ERROR_MESSAGE_PREFIX = "Ошибка при выполении команды: ";
+
     private final BookService bookService;
 
     public UserCommand(BookService bookService) {
@@ -19,15 +18,23 @@ public class UserCommand {
     }
 
     @ShellMethod("Найти книгу")
-    public List<Book> find(
+    public String find(
             @ShellOption(defaultValue = "0") long id,
             @ShellOption(defaultValue = "") String name) {
         if (id > 0) {
-            return Collections.singletonList(bookService.find(id));
+            final var optionalBook = bookService.findById(id);
+            if (optionalBook.isEmpty()) {
+                return ERROR_MESSAGE_PREFIX + "не найдена книга с id " + id;
+            }
+            return optionalBook.get().toString();
         } else if (StringUtils.isNoneEmpty(name)) {
-            return bookService.find(name);
+            final var books = bookService.findByName(name);
+            if (CollectionUtils.isEmpty(books)) {
+                return ERROR_MESSAGE_PREFIX + "не найдены книги с name " + name;
+            }
+            return books.toString();
         }
-        throw new IllegalArgumentException("необходимо задать id или имя книги");
+        return ERROR_MESSAGE_PREFIX + "необходимо задать id или имя книги";
     }
 
     /**
@@ -36,21 +43,32 @@ public class UserCommand {
      * Предполагаем, что у нас один автор и один жанр у книги
      */
     @ShellMethod("Создать книгу")
-    public Book create(String name, String authorName, String genreName) {
-        return bookService.create(name, authorName, genreName);
+    public String create(String name, String authorName, String genreName) {
+        try {
+            return bookService.create(name, authorName, genreName).toString();
+        } catch (Exception ex) {
+            return ERROR_MESSAGE_PREFIX + ex.getMessage();
+        }
     }
 
     @ShellMethod("Обновить книгу")
-    public Book update(
+    public String update(
             long id,
             String name,
             @ShellOption(defaultValue = "") String authorName,
             @ShellOption(defaultValue = "") String genreName) {
-        return bookService.update(id, name, authorName, genreName);
+        try {
+            return bookService.update(id, name, authorName, genreName).toString();
+        } catch (Exception ex) {
+            return ERROR_MESSAGE_PREFIX + ex.getMessage();
+        }
     }
 
     @ShellMethod("Удалить книгу")
-    public String delete(long id) {
+    public String delete(@ShellOption(defaultValue = "0") long id) {
+        if (id == 0) {
+            return ERROR_MESSAGE_PREFIX + "необходимо задать id книги";
+        }
         bookService.delete(id);
         return "done.";
     }
